refactor(CollectionDetail): Promove a página de coleção do usuário a página canônica única (usuário + coletivo).
