+++ src/views/CollectionDetail.vue
@@ -4,9 +4,11 @@ import { useRoute, useRouter } from "vue-router";
 import { useAuthStore } from "@/store/auth";
 import { useAlbumsStore } from "@/store/albums";
 import { useUsersStore } from "@/store/users";
+import { useCollectivesStore } from "@/store/collectives";
 import defaultProfileImage from "@/assets/profile_image.png";
+import collectiveImageDefault from "@/assets/collective_image.png";
 import CollectionPeriodsChart from "@/components/CollectionPeriodsChart.vue";
-import UiField from "../../components/ui/UiField.vue";
+import UiField from "@/components/ui/UiField.vue";
 import CollectionImagesGrid from "@/components/collection/CollectionImagesGrid.vue";
 import CollectionImagesMosaic from "@/components/collection/CollectionImagesMosaic.vue";
 import CollectionImagesMap from "@/components/collection/CollectionImagesMap.vue";
@@ -24,9 +26,12 @@ import { sanitizeDownloadFilename } from "@/helpers/downloadImage";
 defineOptions({ name: "CollectionDetail" });
 
 const usersStore = useUsersStore();
+const collectivesStore = useCollectivesStore();
 const API_BASE_URL = import.meta.env.VITE_BASE_REQUEST_URL;
 
 const ownerUser = ref(null);
+// Dono do tipo coletivo (quando a coleção pertence a um coletivo).
+const collective = ref(null);
 
 const route = useRoute();
 const router = useRouter();
@@ -37,9 +42,17 @@ const isLoadingOwner = ref(true);
 
 const collectionId = computed(() => route.params.collectionId);
 
+// Para visitantes não logados não enviamos header de autorização
+// (authStore.authHeader resolveria para "Bearer null" sem token).
+const userAuthHeader = computed(() =>
+  authStore.isLoggedIn ? authStore.authHeader : null
+);
+
 const albumData = ref(null);
 const collectionImages = ref([]);
 const isLoadingCollection = ref(true);
+// null = ok | "forbidden" | "notfound" | "error"
+const accessState = ref(null);
 
 // título e descrição reais vindos da API
 const collectionTitle = computed(() => {
@@ -54,11 +67,40 @@ const collectionDescription = computed(() => {
 // status de visibilidade da coleção (pública/privada)
 const isPrivate = computed(() => !!albumData.value?.is_private);
 
-// Destino do link do dono da coleção (perfil público). Usa o user_id do álbum,
-// sempre disponível quando a coleção carrega.
+// O dono do álbum é polimórfico: coletivo (collective_id) ou usuário (user_id).
+const isCollectiveOwned = computed(() => !!albumData.value?.collective_id);
+
+// Destino do link do dono da coleção: página do coletivo ou perfil público do usuário.
 const ownerProfileTo = computed(() => {
-  const ownerId = albumData.value?.user_id;
-  return ownerId ? { name: "view-profile", params: { id: ownerId } } : null;
+  if (!albumData.value) return null;
+  if (albumData.value.collective_id) {
+    return { name: "collective-detail", params: { id: albumData.value.collective_id } };
+  }
+  if (albumData.value.user_id) {
+    return { name: "view-profile", params: { id: albumData.value.user_id } };
+  }
+  return null;
+});
+
+// Nome exibido do dono (coletivo ou usuário).
+const ownerName = computed(() => {
+  if (isCollectiveOwned.value) return collective.value?.name?.trim() || "Coletivo";
+  return ownerUser.value?.name?.trim() || "Usuário desconhecido";
+});
+
+// Rótulo da seção do dono.
+const ownerTitle = computed(() =>
+  isCollectiveOwned.value ? "Coleção do coletivo" : "Coleção criada por"
+);
+
+// Quem pode gerenciar (editar): membro do coletivo ou o próprio dono usuário.
+const canManage = computed(() => {
+  const user = authStore.loggedUser;
+  if (!authStore.isLoggedIn || !user || !albumData.value) return false;
+  if (albumData.value.collective_id) {
+    return collective.value?.members?.some((m) => m.id === user.id) ?? false;
+  }
+  return albumData.value.user_id === user.id;
 });
 
 // Navega para a busca filtrada por subject na home, em modo grid:
@@ -72,10 +114,11 @@ function goToSubject(subjectId) {
   });
 }
 
-// avatar do proprietário da coleção
+// avatar do proprietário da coleção (coletivo ou usuário)
 const ownerAvatarSrc = computed(() => {
-  const avatarUrl = ownerUser.value?.avatar_url;
-  const avatarPath = ownerUser.value?.avatar_path;
+  const source = isCollectiveOwned.value ? collective.value : ownerUser.value;
+  const avatarUrl = source?.avatar_url;
+  const avatarPath = source?.avatar_path;
 
   // Caso já venha URL completa da API/CDN
   if (avatarUrl && /^https?:\/\//.test(avatarUrl)) {
@@ -87,63 +130,86 @@ const ownerAvatarSrc = computed(() => {
     return `${API_BASE_URL}${avatarUrl.startsWith("/") ? "" : "/"}${avatarUrl}`;
   }
 
-  // Caso backend retorne avatar_path (apenas pra eu lembra: mesmo padrão do ProfileCard)
+  // Caso backend retorne avatar_path (mesmo padrão do ProfileCard)
   if (avatarPath) {
     return `${API_BASE_URL}/storage/${avatarPath}`;
   }
 
-  return defaultProfileImage;
+  return isCollectiveOwned.value ? collectiveImageDefault : defaultProfileImage;
 
 });
 
-// busca os dados da coleção
+// Busca os dados da coleção, tratando o acesso (403/404). O header de
+// autorização é opcional: visitantes não logados veem coleções públicas.
 async function fetchCollectionData() {
   if (!collectionId.value) return;
 
   isLoadingCollection.value = true;
   isLoadingOwner.value = true;
+  accessState.value = null;
 
-  try {
-    // busca os dados da coleção
-    const data = await albumsStore.getDataAlbumByAlbumId(
-      authStore.authHeader,
-      collectionId.value
-    );
+  const result = await albumsStore.getAlbumDetail(
+    userAuthHeader.value,
+    collectionId.value
+  );
 
-    albumData.value = data;
-
-    // carrega as imagens da coleção
-    await loadCollectionImageDetails(data.images || []);
-
-    // carrega as tags da coleção
-    await loadCollectionTags();
-
-    if (data.user_id) {
-      try {
-        ownerUser.value = await usersStore.getUser(data.user_id);
-      } catch (e) {
-        ownerUser.value = null;
-      } finally {
-        isLoadingOwner.value = false;
-      }
-    } else {
-      ownerUser.value = null;
-      isLoadingOwner.value = false;
-    }
-
-  } catch (error) {
+  if (!result.success) {
     albumData.value = null;
     collectionImages.value = [];
     collectionTags.value = [];
     isLoadingCollectionTags.value = false;
     ownerUser.value = null;
+    collective.value = null;
     isLoadingOwner.value = false;
-
-  } finally {
+    if (result.status === 403) {
+      accessState.value = "forbidden";
+    } else if (result.status === 404) {
+      accessState.value = "notfound";
+    } else {
+      accessState.value = "error";
+    }
     isLoadingCollection.value = false;
-
+    return;
   }
 
+  albumData.value = result.data;
+
+  // carrega as imagens e as tags da coleção
+  await loadCollectionImageDetails(result.data.images || []);
+  await loadCollectionTags();
+  isLoadingCollection.value = false;
+
+  // Resolve o dono conforme o tipo do álbum (coletivo ou usuário).
+  if (result.data.collective_id) {
+    await fetchCollective(result.data.collective_id);
+  } else if (result.data.user_id) {
+    await fetchOwnerUser(result.data.user_id);
+  } else {
+    isLoadingOwner.value = false;
+  }
+}
+
+// Busca os dados do coletivo dono (nome, avatar, membros).
+async function fetchCollective(id) {
+  isLoadingOwner.value = true;
+  try {
+    const result = await collectivesStore.getCollective(id);
+    collective.value = result.success ? result.data : null;
+  } finally {
+    isLoadingOwner.value = false;
+  }
+}
+
+// Busca os dados do usuário dono.
+async function fetchOwnerUser(userId) {
+  isLoadingOwner.value = true;
+  try {
+    ownerUser.value = await usersStore.getUser(userId);
+  } catch (e) {
+    ownerUser.value = null;
+  } finally {
+    isLoadingOwner.value = false;
+  }
 }
 
 // Carrega as imagens da coleção
@@ -169,14 +235,6 @@ async function loadCollectionImageDetails(imagesFromAlbum = []) {
     );
 
     selectedMapImageId.value = null;
-
-    const plainImages = JSON.parse(JSON.stringify(collectionImages.value));
-    console.log("[CollectionDetail] imagens da coleção:", plainImages);
-    console.log("[CollectionDetail] primeira imagem:", plainImages[0]);
-    console.log(
-      "[CollectionDetail] chaves da primeira imagem:",
-      plainImages[0] ? Object.keys(plainImages[0]) : []
-    );
 
   } catch (err) {
     console.error("Erro ao carregar imagens da coleção:", err);
@@ -264,7 +322,7 @@ function handleCollectionViewChange({ selection }) {
   if (viewMode === route.params.viewMode) return;
 
   router.push({
-    name: "my-collection-detail",
+    name: "collection-detail",
     params: {
       collectionId: collectionId.value,
       viewMode,
@@ -388,6 +446,18 @@ async function handleCollectionDownloadConfirm() {
  * End: Download
  */
 
+// "Voltar": retorna à página anterior se houver histórico interno;
+// senão, cai na página do dono (coletivo/perfil) derivada do álbum.
+function goBack() {
+  if (window.history.state?.back) {
+    router.back();
+    return;
+  }
+  const target = ownerProfileTo.value;
+  if (target) router.push(target);
+  else router.push("/");
+}
+
 onMounted(() => {
   fetchCollectionData();
   updateIsMobile();
@@ -413,7 +483,7 @@ watch(
     if (!mobile || viewMode !== "mosaic" || !collectionId.value) return;
 
     router.replace({
-      name: "my-collection-detail",
+      name: "collection-detail",
       params: {
         collectionId: collectionId.value,
         viewMode: "grid",
@@ -433,11 +503,12 @@ watch(
     }"
   >
     <CollectionToolbar
-      v-if="!isMobile"
+      v-if="!isMobile && !accessState"
       class="collection-detail__floating-toolbar"
       :view-selection="viewSelection"
       :is-info-active="isInfoActive"
       :allowed-views="['grid', 'mosaic', 'map']"
+      :can-edit="canManage"
       @view-change="handleCollectionViewChange"
       @toggle-info="handleToggleCollectionInfo"
       @download="handleDownloadCollection"
@@ -455,7 +526,7 @@ watch(
           type="button"
           class="collection-detail__back-btn"
           aria-label="Voltar"
-          @click="router.push({ name: 'my-profile-collections' })"
+          @click="goBack"
         >
             <span class="collection-detail__back-content">
                 <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 14 14" fill="none" aria-hidden="true">
@@ -468,9 +539,37 @@ watch(
                 </svg>
                 <span class="collection-detail__back-text">voltar</span>
             </span>
-        </button>      
+        </button>
     </header>
-    <main class="collection-detail__main container-fluid px-0">
+
+    <!-- Estados de acesso (coleção privada, inexistente ou erro) -->
+    <div
+      v-if="accessState === 'forbidden'"
+      class="collection-detail__state alert alert-dark bg-off-white alert-light border border-dark border-start-3 d-inline-flex align-items-center px-3 py-2"
+      role="status"
+    >
+      <i class="bi bi-lock-fill text-dark me-2"></i>
+      <span>Esta coleção é privada e você não tem permissão para visualizá-la.</span>
+    </div>
+
+    <div
+      v-else-if="accessState === 'notfound'"
+      class="collection-detail__state alert alert-dark bg-off-white alert-light border border-dark border-start-3 d-inline-flex align-items-center px-3 py-2"
+      role="status"
+    >
+      <i class="bi bi-exclamation-circle-fill text-dark me-2"></i>
+      <span>Coleção não encontrada.</span>
+    </div>
+
+    <div
+      v-else-if="accessState === 'error'"
+      class="collection-detail__state collection-detail__state--error"
+      role="alert"
+    >
+      Não foi possível carregar a coleção. Tente novamente.
+    </div>
+
+    <main v-else class="collection-detail__main container-fluid px-0">
       <div
         class="row g-0 collection-detail__row"
         :class="{
@@ -681,7 +780,7 @@ watch(
               <template v-if="!(collectionViewMode === 'map' && selectedMapImage)">
               <section class="collection-detail__actors">
               <div class="collection-detail__actors-title-area">
-                <h2 class="collection-detail__actors-title">Coleção criada por</h2>
+                <h2 class="collection-detail__actors-title">{{ ownerTitle }}</h2>
               </div>
               <div class="collection-detail__actors-list">
                 <div class="collection-detail__actor-image-area">
@@ -693,7 +792,7 @@ watch(
                   <img
                     v-else
                     :src="ownerAvatarSrc"
-                    :alt="`Avatar de ${ownerUser?.name?.trim() || 'usuário'}`"
+                    :alt="`Avatar de ${ownerName}`"
                     class="collection-detail__actor-image"
                   />
                 </div>
@@ -706,7 +805,7 @@ watch(
                     class="collection-detail__actor-name"
                     :class="{ 'collection-detail__actor-name--visible': !isLoadingOwner }"
                   >
-                    {{ ownerUser?.name?.trim() || "Usuário desconhecido" }}
+                    {{ ownerName }}
                   </component>
                 </div>
               </div>
@@ -768,7 +867,7 @@ watch(
             </template>
             </div>
 
-            <div class="collection-detail__mobile-edit">
+            <div v-if="canManage" class="collection-detail__mobile-edit">
               <button
                 type="button"
                 class="collection-detail__mobile-edit-btn"
@@ -860,7 +959,7 @@ watch(
               <template v-if="!(collectionViewMode === 'map' && selectedMapImage)">
               <section class="collection-detail__actors">
                 <div class="collection-detail__actors-title-area">
-                  <h2 class="collection-detail__actors-title">Coleção criada por</h2>
+                  <h2 class="collection-detail__actors-title">{{ ownerTitle }}</h2>
                 </div>
                 <div class="collection-detail__actors-list">
                   <div class="collection-detail__actor-image-area">
@@ -1060,6 +1159,16 @@ a.collection-detail__actor-name {
 
 a.collection-detail__actor-name:hover {
   text-decoration: underline;
+}
+
+.collection-detail__state {
+  margin-top: 8px;
+}
+
+.collection-detail__state--error {
+  color: #7a1c1c;
+  font-family: "DM Sans", sans-serif;
+  font-size: 14px;
 }
 
 .collection-detail__container {
